Return only objects which the user has access to
Update the getinstruments, getProjects, and getstorages apis
to only return the instruments, projects, and storages which
have the user's email in the users assigned array. Update the
models to reflect this users assigned array. Update the
react frontend pages to use post request, with the proper
headings and bodys, and to obtain the user email with
verifyToken.